Flatten zip output before join-ing in full-justify()
The flattening behaviour of some list operators has changed in a recent
version of Rakudo, such that in certain situations the output is no longer
flattened automatically.  When zipping together the words and spaces in
`full-justify()` a list of lists is created instead of a simple flat list.
Thus it is necessary to explicitly flatten this list before joining the
elements together.  This change makes the test suite pass with current
versions of Rakudo.