Pass arrays unflatted into form()
The form() function was expecting an unflatted array to be passed in as its
argument after the format.  This was getting flattened and thus the array of
data was being attempted to be parsed as a form grammar, which meant that
the output was incorrect.  Post GLR one can stop the flattening by prefixing
the array with a dollar sign.